+++ src/components/ripple.ts
@@ -4,101 +4,78 @@ import { useComputedStyle } from '../core/utils/CSS.js'
 import * as scheme from '../core/scheme.js'
 
 const props = useProps({
-  attached: false,
   centered: false
 })
 
 
 const style = /*css*/`
 :host{
-  display: inline-block;
-  vertical-align: middle;
-  position: relative;
-  cursor: pointer;
-}
-:host([attached=true]),
-.container,
-.container::before,
-.ripple{
   position: absolute;
   inset: 0;
   width: 100%;
   height: 100%;
   pointer-events: none;
   border-radius: inherit;
-}
-.container{
   overflow: hidden;
-  &::before{
+  .hover{
+    position: absolute;
+    inset: 0;
     content: '';
     opacity: 0;
     background: var(--ripple-color, currentColor);
     transition: opacity var(--s-motion-duration-short4, ${scheme.motion.duration.short4}) var(--s-motion-easing-standard, ${scheme.motion.easing.standard});
   }
-  &.hovered::before{
+  .hovered{
     opacity: var(--ripple-hover-opacity, .08);
   }
-}
-.ripple{
-  opacity: 0;
-  border-radius: 50%;
-  background: currentColor;
-  filter: opacity(var(--ripple-opacity, .18));
-  animation-timing-function: var(--s-motion-easing-standard, ${scheme.motion.easing.standard});
-  animation-duration: var(--s-motion-duration-long4, ${scheme.motion.duration.long4});
+  .ripple{
+    position: absolute;
+    inset: 0;
+    opacity: 0;
+    border-radius: 50%;
+    background: currentColor;
+    filter: opacity(var(--ripple-opacity, .18));
+    animation-timing-function: var(--s-motion-easing-standard, ${scheme.motion.easing.standard});
+    animation-duration: var(--s-motion-duration-long4, ${scheme.motion.duration.long4});
+  }
 }
 `
 
 const template = /*html*/`
-<slot></slot>
-<div class="container" part="container">
-  <div class="ripple" part="ripple"></div>
-</div>
+<div class="hover" part="hover"></div>
+<div class="ripple" part="ripple"></div>
 `
 
 export class Ripple extends useElement({
   style, template, props,
   setup(shadowRoot) {
-    const container = shadowRoot.querySelector<HTMLDivElement>('.container')!
     const ripple = shadowRoot.querySelector<HTMLDivElement>('.ripple')!
+    const hover = shadowRoot.querySelector<HTMLDivElement>('.hover')!
     const computedStyle = useComputedStyle(ripple)
     const getAnimateOptions = () => {
       const easing = computedStyle.getValue('animation-timing-function')
       const duration = computedStyle.getDuration('animation-duration')
       return { easing, duration }
     }
-    const hover = () => !device.touchEnabled && container.classList.add('hovered')
-    const unHover = () => !device.touchEnabled && container.classList.remove('hovered')
-    const state = { parentNode: null as null | HTMLElement, pressed: false }
-    const run = (event: PointerEvent) => {
-      const { offsetWidth, offsetHeight } = this
-      let size = Math.sqrt(offsetWidth ** 2 + offsetHeight ** 2)
+    const run = (event: PointerEvent, parent: HTMLElement) => {
+      let size = Math.sqrt(parent.offsetWidth ** 2 + parent.offsetHeight ** 2)
       const coordinate = { x: '50%', y: '50%' }
       if (!this.centered) {
-        const { left, top } = this.getBoundingClientRect()
-        const x = event.clientX - left
-        const y = event.clientY - top
-        const h = offsetHeight / 2
-        const w = offsetWidth / 2
-        const edgeW = (Math.abs(h - y) + h) * 2
-        const edgeH = (Math.abs(w - x) + w) * 2
+        const { left, top } = parent.getBoundingClientRect()
+        const state = { x: event.clientX - left, y: event.clientY - top, h: parent.offsetHeight / 2, w: parent.offsetWidth / 2 }
+        const edgeW = (Math.abs(state.h - state.y) + state.h) * 2
+        const edgeH = (Math.abs(state.w - state.x) + state.w) * 2
         size = Math.sqrt(edgeW ** 2 + edgeH ** 2)
-        coordinate.x = `${x}px`
-        coordinate.y = `${y}px`
+        coordinate.x = `${state.x}px`
+        coordinate.y = `${state.y}px`
       }
       let newRipple = ripple
-      let callback = () => { }
-      if (state.pressed) {
+      if (newRipple.getAnimations().length > 0) {
         newRipple = ripple.cloneNode() as HTMLDivElement
-        container.appendChild(newRipple)
-        callback = () => newRipple.remove()
-      } else {
-        state.pressed = true
-        callback = () => state.pressed = false
+        shadowRoot.appendChild(newRipple)
       }
-      const parent = (state.parentNode ?? this)
       const animateOptions = getAnimateOptions()
-      parent.setAttribute('pressed', '')
+      parent.setAttribute('ripple-pressed', '')
       const animation = newRipple.animate({
         opacity: [1, 1],
         width: [`${size}px`, `${size}px`],
@@ -108,19 +85,20 @@ export class Ripple extends useElement({
         top: [coordinate.y, coordinate.y],
       }, { ...animateOptions, fill: 'forwards' })
       return () => {
-        parent.removeAttribute('pressed')
+        parent.removeAttribute('ripple-pressed')
         const time = Number(animation.currentTime)
         const short = animateOptions.duration / 2
         const diff = animateOptions.duration - short
         const duration = time > diff ? short : animateOptions.duration - time
-        newRipple.animate({ opacity: [1, 0] }, { duration, easing: animateOptions.easing, fill: 'forwards' }).finished.then(callback)
+        const animate = newRipple.animate({ opacity: [1, 0] }, { ...animateOptions, duration, easing: animateOptions.easing, fill: 'forwards' })
+        animate.finished.then(() => newRipple !== ripple && shadowRoot.removeChild(newRipple))
       }
     }
-    const down = async (event: PointerEvent) => {
+    function down(this: HTMLElement, event: PointerEvent) {
       if (event.button !== 0) return
-      if (event.pointerType !== 'touch') return document.addEventListener('pointerup', run(event), { once: true })
+      if (event.pointerType !== 'touch') return document.addEventListener('pointerup', run(event, this), { once: true })
       let stop: Function
-      const timer = setTimeout(() => stop = run(event), 40)
+      const timer = setTimeout(() => stop = run(event, this), 40)
       const move = () => {
         if (stop) return document.removeEventListener('touchmove', move)
         clearTimeout(timer)
@@ -133,34 +111,28 @@ export class Ripple extends useElement({
       document.addEventListener('touchmove', move, { passive: true })
       document.addEventListener('touchend', remove, { passive: true })
     }
-    const add = (target: HTMLElement) => {
-      target.addEventListener('mouseenter', hover)
-      target.addEventListener('mouseleave', unHover)
-      target.addEventListener('wheel', unHover, { passive: true })
-      target.addEventListener('pointerdown', down)
+    function hovering(this: HTMLElement, event: MouseEvent) {
+      if (device.touchEnabled) return
+      this.setAttribute('ripple-hovered', '')
+      hover.classList.add('hovered')
     }
-    const remove = () => {
-      if (!state.parentNode) return
-      state.parentNode.removeEventListener('mouseenter', hover)
-      state.parentNode.removeEventListener('mouseleave', unHover)
-      state.parentNode.removeEventListener('wheel', unHover)
-      state.parentNode.removeEventListener('pointerdown', down)
-      state.parentNode = null
+    function unHovering(this: HTMLElement, event: MouseEvent) {
+      if (device.touchEnabled) return
+      this.removeAttribute('ripple-hovered')
+      hover.classList.remove('hovered')
     }
-    add(this)
     return {
-      onMounted: () => {
-        if (this.attached && this.parentNode) {
-          state.parentNode = (this.parentNode instanceof ShadowRoot ? this.parentNode.host : this.parentNode) as HTMLElement
-          add(state.parentNode)
+      onMounted: (parent) => {
+        const parentElement = parent instanceof ShadowRoot ? parent.host : parent
+        if (!(parentElement instanceof HTMLElement)) return
+        parentElement.addEventListener('pointerdown', down)
+        parentElement.addEventListener('mouseenter', hovering)
+        parentElement.addEventListener('mouseleave', unHovering)
+        return () => {
+          parentElement.removeEventListener('pointerdown', down)
+          parentElement.removeEventListener('mouseenter', hovering)
+          parentElement.removeEventListener('mouseleave', unHovering)
         }
-      },
-      onUnmounted: () => this.attached && remove(),
-      setAttached: (value) => {
-        if (!this.isConnected) return
-        if (!value) return remove()
-        const target = (this.parentNode instanceof ShadowRoot ? this.parentNode.host : this.parentNode) as HTMLElement
-        add(target)
       }
     }
   }