+++ src/components/tooltip.ts
@@ -5,146 +5,174 @@ import { useComputedStyle } from '../core/utils/CSS.js'
 import * as scheme from '../core/scheme.js'
 
 const props = useProps({
-  align: ['top', 'bottom', 'left', 'right']
+  placement: ['bottom', 'top', 'left', 'right']
 })
+const events = {
+  opened: Event,
+  closed: Event
+}
 
 const style = /*css*/`
 :host{
-  display: inline-block;
-  vertical-align: middle;
-  text-align: left;
-}
-.popup{
-  position: fixed;
-  display: none;
-  inset: 0;
-  margin: 0;
-  width: fit-content;
-  height: fit-content;
-  background: none;
-  border: none;
-  outline: none;
-  max-width: none;
-  max-height: none;
-  font-size: .875rem;
-  font-weight: 400;
-  padding: 6px 8px;
-  border-radius: 4px;
-  white-space: nowrap;
-  filter: opacity(.85);
-  background: var(--s-color-inverse-surface, ${scheme.color.inverseSurface});
-  color: var(--s-color-inverse-on-surface, ${scheme.color.inverseOnSurface});
+  position: absolute;
   animation-timing-function: var(--s-motion-easing-standard, ${scheme.motion.easing.standard});
   animation-duration: var(--s-motion-duration-medium4, ${scheme.motion.duration.medium4});
-}
-::slotted(img){
-  display: block;
-  border-radius: 4px;
-  margin: 2px 0;
-  max-width: 280px;
+  --tooltip-placement: none;
+  .popover{
+    pointer-events: none;
+    position: fixed;
+    display: none;
+    inset: 0;
+    margin: 0;
+    width: fit-content;
+    height: fit-content;
+    background: none;
+    border: none;
+    overflow: hidden;
+    outline: none;
+    max-width: 100%;
+    max-height: 100%;
+    font-size: .875rem;
+    font-weight: 400;
+    padding: 6px 8px;
+    border-radius: 4px;
+    filter: opacity(.95);
+    background: var(--s-color-inverse-surface, ${scheme.color.inverseSurface});
+    color: var(--s-color-inverse-on-surface, ${scheme.color.inverseOnSurface});
+  }
+  .opened{
+    display: block;
+  }
 }
 `
 
 const template = /*html*/`
-<slot name="trigger" part="trigger"></slot>
-<div class="popup" popover="manual" part="popup">
-  <slot></slot>
-</div>
+<slot class="popover" popover="manual" part="popup"></slot>
 `
 
 export class Tooltip extends useElement({
-  style, template, props,
+  style, template, props, events,
   setup(shadowRoot) {
-    const trigger = shadowRoot.querySelector<HTMLSlotElement>('slot[name=trigger]')!
-    const popup = shadowRoot.querySelector<HTMLDivElement>('.popup')!
-    const computedStyle = useComputedStyle(popup)
-    let showed = false
+    const popover = shadowRoot.querySelector<HTMLSlotElement>('.popover')!
+    const computedStyle = useComputedStyle(this)
     const getAnimateOptions = () => {
       const easing = computedStyle.getValue('animation-timing-function')
       const duration = computedStyle.getDuration('animation-duration')
       return { easing, duration }
     }
-    const show = () => {
-      if (!this.isConnected) return
-      popup.style.display = 'block'
-      showed = true
-      if (popup.showPopover) {
-        popup.showPopover()
-      } else {
-        const rect = getStackingContext(shadowRoot)
-        popup.style.marginLeft = `${-rect.left}px`
-        popup.style.marginTop = `${-rect.top}px`
-        popup.style.zIndex = '3'
-      }
-      const rect = this.getBoundingClientRect()
+    const display = ['block', 'block']
+    const show = (parent: HTMLElement) => {
+      if (!this.isConnected || popover.classList.contains('opened')) return
+      if (!this.dispatchEvent(new Event('open', { cancelable: true }))) return
+      popover.style.display = 'block'
       const gap = 4
-      const cWidth = popup.offsetWidth
-      const cHeight = popup.offsetHeight
+      popover.style.removeProperty('top')
+      popover.style.removeProperty('left')
+      const { offsetHeight, offsetWidth } = popover
+      if (!popover.showPopover) {
+        const rect = getStackingContext(shadowRoot)
+        popover.style.marginLeft = `${-rect.left}px`
+        popover.style.marginTop = `${-rect.top}px`
+        popover.style.zIndex = '3'
+      }
+      popover.style.removeProperty('display')
+      popover?.showPopover()
+      const rect = parent.getBoundingClientRect()
+      const offsets = {
+        top: rect.top - gap - offsetHeight,
+        bottom: rect.top + gap + rect.height,
+        left: rect.left - gap - offsetWidth,
+        right: rect.left + gap + rect.width
+      }
+      const vertical = { name: 'top', side: { name: 'left', size: offsetWidth, innerSize: innerWidth, targetSize: rect.width } } as const
+      const horizontal = { name: 'left', side: { name: 'top', size: offsetHeight, innerSize: innerHeight, targetSize: rect.height } } as const
+      const options = {
+        top: { value: offsets.top, overflowed: offsets.top < 0, ...vertical },
+        bottom: { value: offsets.bottom, overflowed: offsets.bottom + offsetHeight > innerHeight, ...vertical },
+        left: { value: offsets.left, overflowed: offsets.left < 0, not: 'right', ...horizontal },
+        right: { value: offsets.right, overflowed: offsets.right + offsetWidth > innerWidth, ...horizontal }
+      } as const
       const position = { top: 0, left: 0 }
-      const calls = {
-        middle(align: 'top' | 'bottom') {
-          position.left = rect.left - ((cWidth - rect.width) / 2)
-          const bottom = () => {
-            position.top = rect.top + rect.height + gap
-            return position.top + cHeight > innerHeight
+      if (options.top.overflowed && options.bottom.overflowed && options.left.overflowed && options.right.overflowed) {
+        position.top = (innerHeight - offsetHeight) / 2
+        position.left = (innerWidth - offsetWidth) / 2
+      } else {
+        const aligns = {
+          top: ['bottom', 'left', 'right'],
+          bottom: ['top', 'left', 'right'],
+          left: ['right', 'top', 'bottom'],
+          right: ['left', 'top', 'bottom']
+        }
+        const value = computedStyle.getValue('--tooltip-placement')
+        const placement = (value in aligns ? value : this.placement) as keyof typeof aligns
+        for (const key of [placement, ...aligns[placement]]) {
+          const name = key as keyof typeof options
+          const option = options[name]
+          if (!option.overflowed) {
+            position[option.name] = option.value
+            const side = option.side
+            position[side.name] = rect[side.name] - (side.size - side.targetSize) / 2
+            if (position[side.name] < 0 && rect[side.name] + side.size < side.innerSize) {
+              position[side.name] = rect[side.name]
+              break
+            }
+            if (position[side.name] + side.size > side.innerSize && rect[side.name] - side.size > 0) {
+              position[side.name] = rect[side.name] + side.targetSize - side.size
+              break
+            }
+            if (position[side.name] < 0 || position[side.name] + side.size > side.innerSize) position[side.name] = 0
+            break
           }
-          const top = () => {
-            position.top = rect.top - cHeight - gap
-            return position.top < 0
-          }
-          if (position.left < 0) position.left = rect.left
-          if (position.left + cWidth > innerWidth) position.left = rect.left + rect.width - cWidth
-          if (align === 'top') top() && bottom()
-          if (align === 'bottom') bottom() && top()
-        },
-        left() {
-          position.left = rect.left - cWidth - gap
-          position.top = rect.top - ((cHeight - rect.height) / 2)
-          return position.left < 0
-        },
-        right() {
-          position.left = rect.left + rect.width + gap
-          position.top = rect.top - ((cHeight - rect.height) / 2)
-          return position.left + cWidth > innerWidth
         }
       }
-      switch (this.align) {
-        case 'bottom':
-        case 'top':
-          calls.middle(this.align)
-          break
-        case 'left':
-          calls.left() && calls.right()
-          break
-        case 'right':
-          calls.right() && calls.left()
-          break
-      }
-      popup.style.top = `${position.top}px`
-      popup.style.left = `${position.left}px`
-      popup.animate({ opacity: [0, 1] }, getAnimateOptions())
+      popover.style.top = `${position.top}px`
+      popover.style.left = `${position.left}px`
+      popover.classList.add('opened')
+      parent.setAttribute('tooltip-opened', '')
+      popover.animate({ opacity: [0, 1], display }, getAnimateOptions()).finished.then(() => this.dispatchEvent(new Event('opened')))
+      this.dispatchEvent(new Event('s-top-layer-open', { bubbles: true }))
     }
-    let timer = 0
-    const close = () => {
-      clearTimeout(timer)
-      showed = false
-      const animation = popup.animate({ opacity: [1, 0] }, getAnimateOptions())
+    const close = (parent: HTMLElement) => {
+      if (!this.isConnected || !popover.classList.contains('opened')) return
+      if (!this.dispatchEvent(new Event('close', { cancelable: true }))) return
+      popover.classList.remove('opened')
+      parent.removeAttribute('tooltip-opened')
+      const animation = popover.animate({ opacity: [1, 0], display }, getAnimateOptions())
       animation.finished.then(() => {
-        if (showed) return
-        popup.hidePopover && popup.hidePopover()
-        popup.style.removeProperty('display')
+        if (popover.classList.contains('opened')) return
+        popover?.hidePopover()
+        this.dispatchEvent(new Event('closed'))
       })
     }
-    trigger.addEventListener('touchstart', () => {
-      clearTimeout(timer)
-      timer = setTimeout(() => show(), 600)
-    }, { passive: true })
-    trigger.ontouchend = () => {
-      clearTimeout(timer)
-      close()
+    return {
+      onMounted: (parent) => {
+        const parentElement = parent instanceof ShadowRoot ? parent.host : parent
+        if (!(parentElement instanceof HTMLElement)) return
+        const hover = () => !device.touchEnabled && show(parentElement)
+        const unHover = () => !device.touchEnabled && close(parentElement)
+        let timer: number
+        const touchstart = () => {
+          console.log('touchstart')
+          timer = setTimeout(() => show(parentElement), 500)
+        }
+        const touchend = () => {
+          clearTimeout(timer)
+          close(parentElement)
+        }
+        parentElement.addEventListener('mouseenter', hover)
+        parentElement.addEventListener('mouseleave', unHover)
+        parentElement.addEventListener('wheel', unHover)
+        parentElement.addEventListener('touchstart', touchstart, { passive: true })
+        parentElement.addEventListener('touchend', touchend)
+        return () => {
+          parentElement.removeEventListener('mouseenter', hover)
+          parentElement.removeEventListener('mouseleave', unHover)
+          parentElement.removeEventListener('wheel', unHover)
+          parentElement.removeEventListener('touchstart', touchstart)
+          parentElement.removeEventListener('touchend', touchend)
+        }
+      }
     }
-    trigger.onmouseenter = () => !device.touchEnabled && show()
-    trigger.onmouseleave = () => !device.touchEnabled && close()
   }
 }) { }
 
@@ -158,7 +186,7 @@ declare global {
     namespace JSX {
       interface IntrinsicElements {
         //@ts-ignore
-        [name]: React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement>, HTMLElement> & Partial<typeof props>
+        [name]: React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement>, HTMLElement>
       }
     }
   }
@@ -173,7 +201,7 @@ declare module 'vue' {
       /**
       * @deprecated
       **/
-      $props: HTMLAttributes & Partial<typeof props>
+      $props: HTMLAttributes
     } & Tooltip
   }
 }
@@ -183,7 +211,7 @@ declare module 'vue/jsx-runtime' {
   namespace JSX {
     export interface IntrinsicElements {
       //@ts-ignore
-      [name]: IntrinsicElements['div'] & Partial<typeof props>
+      [name]: IntrinsicElements['div']
     }
   }
 }
@@ -193,7 +221,7 @@ declare module 'solid-js' {
   namespace JSX {
     interface IntrinsicElements {
       //@ts-ignore
-      [name]: JSX.HTMLAttributes<HTMLElement> & Partial<typeof props>
+      [name]: JSX.HTMLAttributes<HTMLElement>
     }
   }
 }
@@ -203,7 +231,7 @@ declare module 'preact' {
   namespace JSX {
     interface IntrinsicElements {
       //@ts-ignore
-      [name]: JSXInternal.HTMLAttributes<HTMLElement> & Partial<typeof props>
+      [name]: JSXInternal.HTMLAttributes<HTMLElement>
     }
   }
 }